requirements.txt
python-dotenv==1.0.0
openai-agents>=0.2.8
tavily-python>=0.7.11